Lifestoned.DataModel converted to .net standard 2.0
This switches two dependancies to ConfigurationManager to functions that accept arguments from the caller (that would query Configurationmanager)
- WorldRelease\Version.cs
- WorldRelease\WorldRelease.cs

Basically, in both of these cases, DerethForever.Web will query the ConfigurationManager and pass the result to Lifestoned.DataModel.

This allows Lifestoned.DatModel to exist on it's own, decoupled from the parent.